+++ docsig/_message.py
@@ -1,0 +1,30 @@
+"""
+docsig._message
+===============
+"""
+from __future__ import annotations as _
+
+import typing as _t
+
+
+class Message(_t.NamedTuple):
+    """Message type for errors."""
+
+    code: str
+    description: str
+    hint: _t.Optional[str] = None
+
+    def fstring(self, template: str) -> str:
+        """Return values as a format string.
+
+        :param template: String to interpolate values.
+        :return: Formatted string.
+        """
+        return template.format(
+            code=self.code,
+            description=self.description,
+        )
+
+
+class Messages(_t.Dict[int, Message]):
+    """Object for storing and retrieving message objects."""